backend/requirements.txt
Django>=2.1.3
djangorestframework>=3.9.0
django-environ
django-filter
django-cors-headers
python-dotenv
Pillow
requests
drf_yasg
torch
torchvision
timm